Cleanup, Renaming - Rename SidebarViews - Update listsLoad to use `first` for `include_bookmarks`
