+++ frontend/src/components/controls/SidebarListViewerUI/style.scss
