+++ frontend/src/components/controls/SidebarPageViewerUI/style.scss
