+++ frontend/src/components/controls/index.js
@@ -16,8 +16,8 @@ export ReplayPageDisplay from './ReplayPageDisplay';
 export ReplayUI from './ReplayUI';
 export ReplayURLBar from './ReplayURLBar';
 export ShareWidgetUI from './ShareWidgetUI';
-export SidebarListViewer from './SidebarListViewer';
-export SidebarPageViewer from './SidebarPageViewer';
+export SidebarListViewerUI from './SidebarListViewerUI';
+export SidebarPageViewerUI from './SidebarPageViewerUI';
 export SidebarUI from './SidebarUI';
 export SizeCounter from './SizeCounter';
 export StandaloneRecorderUI from './StandaloneRecorderUI';
