+++ frontend/src/components/icons/Info.js
@@ -3,15 +3,10 @@ import React from 'react';
 
 function Info() {
   return (
-    <svg width="22px" height="22px" viewBox="0 0 22 22" version="1.1" xmlns="http://www.w3.org/2000/svg">
-      <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-        <g id="00_Launch-Screen" transform="translate(-1236.000000, -69.000000)">
-          <g id="info_ico" transform="translate(1237.000000, 70.000000)">
-            <path d="M12.3491379,14.0933908 L12.191092,14.7399425 C11.7169517,14.9267251 11.3386029,15.0692045 11.0560345,15.1673851 C10.773466,15.2655656 10.4454042,15.3146552 10.0718391,15.3146552 C9.49712356,15.3146552 9.05052841,15.1745704 8.73204023,14.8943966 C8.41355205,14.6142227 8.25431034,14.2586228 8.25431034,13.8275862 C8.25431034,13.6599608 8.26628341,13.4887461 8.29022989,13.3139368 C8.31417636,13.1391275 8.35249016,12.941572 8.40517241,12.7212644 L8.99425287,10.6235632 C9.04693513,10.4224128 9.09123545,10.2320411 9.12715517,10.0524425 C9.16307489,9.87284393 9.18103448,9.708813 9.18103448,9.56034483 C9.18103448,9.29214425 9.12595841,9.10536451 9.0158046,9 C8.90565079,8.89463549 8.69492493,8.84195402 8.38362069,8.84195402 C8.23036322,8.84195402 8.07351613,8.86590014 7.91307471,8.9137931 C7.7526333,8.96168606 7.6149431,9.00718369 7.5,9.05028736 L7.65804598,8.40373563 C8.04597895,8.24568886 8.41714382,8.11039328 8.77155172,7.99784483 C9.12595963,7.88529637 9.46120532,7.82902299 9.77729885,7.82902299 C10.3472251,7.82902299 10.7866364,7.96671318 11.095546,8.2420977 C11.4044556,8.51748222 11.558908,8.87547672 11.558908,9.31609195 C11.558908,9.40708858 11.5481323,9.56752759 11.5265805,9.79741379 C11.5050286,10.0273 11.4655175,10.2380259 11.408046,10.4295977 L10.8189655,12.5201149 C10.7710726,12.6877403 10.7279695,12.8793093 10.6896552,13.0948276 C10.6513408,13.3103459 10.6321839,13.4731795 10.6321839,13.5833333 C10.6321839,13.8611125 10.6944438,14.0502869 10.8189655,14.1508621 C10.9434872,14.2514373 11.1590023,14.3017241 11.4655172,14.3017241 C11.6091961,14.3017241 11.7720297,14.2765807 11.954023,14.2262931 C12.1360162,14.1760055 12.2677199,14.1317052 12.3491379,14.0933908 Z M12.5,5.32183908 C12.5,5.68582557 12.3623098,5.99592784 12.0869253,6.25215517 C11.8115408,6.50838251 11.479887,6.63649425 11.091954,6.63649425 C10.704021,6.63649425 10.37117,6.50838251 10.0933908,6.25215517 C9.81561164,5.99592784 9.67672414,5.68582557 9.67672414,5.32183908 C9.67672414,4.95785259 9.81561164,4.64655302 10.0933908,4.38793103 C10.37117,4.12930905 10.704021,4 11.091954,4 C11.479887,4 11.8115408,4.12930905 12.0869253,4.38793103 C12.3623098,4.64655302 12.5,4.95785259 12.5,5.32183908 Z" id="i" fill="#000000" />
-            <g id="more-info" strokeLinecap="square" stroke="#000000">
-              <circle id="Oval-5" cx="10" cy="10" r="10" />
-            </g>
-          </g>
+    <svg width="16px" height="16px" viewBox="0 0 34 34" version="1.1" xmlns="http://www.w3.org/2000/svg">
+      <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g transform="translate(-32.000000, -31.000000)" fill="#000">
+          <path d="M48.9999595,31.7813197 C51.9660986,31.7813197 54.7108834,32.5228547 57.2343148,34.0059237 C59.7577463,35.4889928 61.7609971,37.4922435 63.2440661,40.015675 C64.7271352,42.5391065 65.4686702,45.2838913 65.4686702,48.2500304 C65.4686702,51.2161695 64.7271352,53.9609543 63.2440661,56.4843858 C61.7609971,59.0078172 59.7577463,61.011068 57.2343148,62.4941371 C54.7108834,63.9772061 51.9660986,64.7187411 48.9999595,64.7187411 C46.0338204,64.7187411 43.2890356,63.9772061 40.7656041,62.4941371 C38.2421726,61.011068 36.2389218,59.0078172 34.7558528,56.4843858 C33.2727838,53.9609543 32.5312487,51.2161695 32.5312487,48.2500304 C32.5312487,45.2838913 33.2727838,42.5391065 34.7558528,40.015675 C36.2389218,37.4922435 38.2421726,35.4889928 40.7656041,34.0059237 C43.2890356,32.5228547 46.0338204,31.7813197 48.9999595,31.7813197 Z M48.9999595,39.0859897 C48.2473574,39.0859897 47.5943635,39.3626821 47.0409798,39.9160659 C46.487596,40.4694497 46.2109036,41.1224436 46.2109036,41.8750456 C46.2109036,42.6276476 46.487596,43.2806415 47.0409798,43.8340253 C47.5943635,44.3874091 48.2473574,44.6641014 48.9999595,44.6641014 C49.7525615,44.6641014 50.4055554,44.3874091 50.9589392,43.8340253 C51.5123229,43.2806415 51.7890153,42.6276476 51.7890153,41.8750456 C51.7890153,41.1224436 51.5123229,40.4694497 50.9589392,39.9160659 C50.4055554,39.3626821 49.7525615,39.0859897 48.9999595,39.0859897 Z M52.7187006,55.953137 L52.7187006,54.3593908 C52.7187006,54.1380375 52.6412265,53.9498863 52.4862793,53.7949391 C52.3313321,53.6399918 52.1431808,53.5625177 51.9218275,53.5625177 L51.1249544,53.5625177 L51.1249544,46.9219086 C51.1249544,46.7005553 51.0474803,46.512404 50.8925331,46.3574568 C50.7375859,46.2025096 50.5494346,46.1250355 50.3280813,46.1250355 L46.0780914,46.1250355 C45.8567381,46.1250355 45.6685869,46.2025096 45.5136397,46.3574568 C45.3586924,46.512404 45.2812183,46.7005553 45.2812183,46.9219086 L45.2812183,48.5156548 C45.2812183,48.7370081 45.3586924,48.9251593 45.5136397,49.0801065 C45.6685869,49.2350538 45.8567381,49.3125279 46.0780914,49.3125279 L46.8749645,49.3125279 L46.8749645,53.5625177 L46.0780914,53.5625177 C45.8567381,53.5625177 45.6685869,53.6399918 45.5136397,53.7949391 C45.3586924,53.9498863 45.2812183,54.1380375 45.2812183,54.3593908 L45.2812183,55.953137 C45.2812183,56.1744903 45.3586924,56.3626416 45.5136397,56.5175888 C45.6685869,56.672536 45.8567381,56.7500101 46.0780914,56.7500101 L51.9218275,56.7500101 C52.1431808,56.7500101 52.3313321,56.672536 52.4862793,56.5175888 C52.6412265,56.3626416 52.7187006,56.1744903 52.7187006,55.953137 Z" />
         </g>
       </g>
     </svg>
